test(downloads): split downloads coverage by package
Why:
- Downloads coverage still ran in the app-host bundle despite exercising package-owned presentation, indexing, repository, and reader contracts.
- The migration plan calls for moving package logic tests out of AndBibleTests so CI can run them in the faster package lanes.

What Changed:
- Moved Downloads presentation/default-language/filter/error contracts into BibleUITests.
- Moved queued search-index deletion and MyBible reader payload readability contracts into BibleCoreTests.
- Moved SWORD and MyBible repository refresh, install layout, rollback, ZIP import, package fallback, and cancellation coverage into SwordKitTests.
- Added the CLibSword dependency needed by SwordKitTests gzip fixtures and removed the old app-host Downloads source from the Xcode project.
- Updated the migration plan with the Downloads ownership split and the intentionally retained SwordManager package coverage.

Validation:
- python3 scripts/check_repo_standards.py docblocks --all-files
- git diff --cached --check
- plutil -lint AndBible.xcodeproj/project.pbxproj
- swift package describe --type json
- DEVELOPER_DIR=/Applications/Xcode.app/Contents/Developer xcodebuild test -project AndBible.xcodeproj -scheme BibleCoreTests -configuration Debug -destination 'platform=iOS Simulator,name=iPhone 16,OS=18.6' -derivedDataPath .derivedData/LocalBibleCorePackageTests CODE_SIGNING_ALLOWED=NO -only-testing:BibleCoreTests/SearchIndexServiceMutationTests -only-testing:BibleCoreTests/MyBibleReaderPackageTests
- DEVELOPER_DIR=/Applications/Xcode.app/Contents/Developer xcodebuild test -project AndBible.xcodeproj -scheme SwordKitTests -configuration Debug -destination 'platform=iOS Simulator,name=iPhone 16,OS=18.6' -derivedDataPath .derivedData/LocalSwordKitPackageTests CODE_SIGNING_ALLOWED=NO -only-testing:SwordKitTests/ModuleRepositoryDownloadTests
- DEVELOPER_DIR=/Applications/Xcode.app/Contents/Developer xcodebuild test -project AndBible.xcodeproj -scheme BibleUITests -configuration Debug -destination 'platform=iOS Simulator,name=iPhone 16,OS=18.6' -derivedDataPath .derivedData/LocalBibleUIPackageTests CODE_SIGNING_ALLOWED=NO -only-testing:BibleUITests/ModuleBrowserDownloadsTests
- DEVELOPER_DIR=/Applications/Xcode.app/Contents/Developer xcodebuild build-for-testing -project AndBible.xcodeproj -scheme AndBible -configuration Debug -destination 'platform=iOS Simulator,name=iPhone 16,OS=18.6' -derivedDataPath .derivedData/LocalAndBibleBuildForTesting CODE_SIGNING_ALLOWED=NO
- npx gitnexus analyze
- gitnexus detect_changes staged reported no affected execution flows

Impact:
- Reduces app-host test bundle scope without changing production behavior.
- Keeps Android-compatible Downloads, MyBible, repository, and cancellation contracts in the owning package test targets.
