+++ docs/test-architecture-migration-plan.md
@@ -149,6 +149,7 @@ Blocker: `AndBibleTestSupport.swift` is a 2,233-line `extension AndBibleTests` e
 - `AndBibleTests+RemoteSyncBookmarks` has moved to `BibleCoreTests` because it exercises Android-compatible bookmark snapshot restore, sparse patch replay/upload, initial backup handling, and bookmark category synchronization through BibleCore services without UI or app bootstrap.
 - `AndBibleTests+ExternalDocumentImport` has been split by owning module: document import routing, Android toast feedback, provider filename normalization, and installer selection moved to `BibleUITests`, while Android-style TTF font repository filesystem contracts moved to `SwordKitTests`.
 - `AndBibleTests+WorkspaceAndRepository` has been split by owning module: workspace settings/store/window-manager/selection, bookmark-label filtering, and global text-display propagation moved to `BibleCoreTests`; the UI-owned workspace text-options dirty-field propagation helper moved to `BibleUITests`; repository-source-manager manifest, sidecar, default-source, and reset contracts moved to `SwordKitTests`.
+- `AndBibleTests+Downloads` has been split by owning module: Downloads browser presentation, Android filter/default-language, auto-refresh, cancellation, and localized error-copy contracts moved to `BibleUITests`; queued search-index deletion and MyBible reader payload readability moved to `BibleCoreTests`; SWORD/MyBible repository refresh, install layout, rollback, package fallback, ZIP import, and cancellation contracts moved to `SwordKitTests`. The duplicated restored-MyBible `SwordManager.installedModules()` cases were not re-migrated because equivalent package coverage already lives in `SwordManagerTests`.
 - Batches for the genuinely BibleUI-behavior remainder: ReaderNavigation, Bookmarks/Strongs, Window/Settings, view-model parts of `+AppAndReader`.
 - **Split `+AppAndReader` deliberately**: `sceneConfiguration` test stays app-hosted; `testContentViewDoesNotContainLegacyRootSidebarShell` becomes a repo-standards guardrail or stays app-hosted (decide explicitly - do not move to BibleUITests); the other ~78 funcs go to BibleUITests.
 - Run BibleUITests via `xcodebuild test -scheme BibleUITests -destination 'platform=iOS Simulator,...'` against the **committed package test scheme from Phase 0** (no app host).
